requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
discord.py==1.7.3
idna==3.3
multidict==5.2.0
requests==2.26.0
rocket-python==1.2.17
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
